Split integration test harness into common/ and clickpipes/
`tests/integration/` previously mixed generic infra (TestContext,
CleanupRegistry, ClickHouse provisioning) with ClickPipes-specific
helpers (AWS, EC2, Kinesis, Redpanda). With the ClickPipes E2E suite
grown to nine binaries, the shared module had become large and the
boundary fuzzy.

- `tests/common/support.rs` — generic infra, used by every integration
  binary (cloud-service CRUD, Postgres-service CRUD, all ClickPipes).
- `tests/clickpipes/support.rs` — ClickPipes-specific helpers;
  `pub use`-re-exports `crate::common::support::*` so callers still
  get both surfaces from `use crate::support::*;`.
- Per-source binaries move from `tests/integration_clickpipe_<src>_test.rs`
  to `tests/clickpipes/<src>_test.rs` and are now named
  `clickpipe_<src>_test`. Cargo doesn't auto-discover `.rs` files in
  subdirectories of `tests/`, so each is declared as an explicit
  `[[test]]` entry in Cargo.toml.
- CI workflow, cloud-api README, and CLAUDE.md updated to reflect the
  new binary names and layout.
