+++ crates/clickhouse-cloud-api/tests/clickpipes/mongo_test.rs
@@ -1,10 +1,14 @@
 //! MongoDB ClickPipe E2E — own CHC service, single stage.
 
-mod integration;
+#[path = "../common/mod.rs"]
+mod common;
+mod driver;
+mod stages;
+mod support;
 
-use integration::driver::*;
-use integration::stages::*;
-use integration::support::*;
+use driver::*;
+use stages::*;
+use support::*;
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 4)]
 #[ignore = "requires live ClickHouse Cloud + AWS credentials and provisions real resources"]
