+++ crates/clickhouse-cloud-api/tests/clickpipes/postgres_ec2_test.rs
@@ -1,13 +1,17 @@
 //! Postgres-on-EC2 ClickPipe E2E — own CHC service, single stage. Parallels
 //! the mysql/mongo per-source binaries. Distinct from the existing
-//! `integration_clickpipe_postgres_cdc_test` (Al's library-driven test
-//! against CHC-managed Postgres).
+//! `clickpipe_postgres_cdc_test` (Al's library-driven test against
+//! CHC-managed Postgres).
 
-mod integration;
+#[path = "../common/mod.rs"]
+mod common;
+mod driver;
+mod stages;
+mod support;
 
-use integration::driver::*;
-use integration::stages::*;
-use integration::support::*;
+use driver::*;
+use stages::*;
+use support::*;
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 4)]
 #[ignore = "requires live ClickHouse Cloud + AWS credentials and provisions real resources"]
