+++ crates/clickhouse-cloud-api/tests/clickpipes/stages/mongo_user_data.sh.template
