+++ crates/clickhouse-cloud-api/tests/clickpipes/stages/mysql_user_data.sh.template
