+++ crates/clickhouse-cloud-api/tests/clickpipes/stages/postgres_user_data.sh.template
