+++ crates/clickhouse-cloud-api/tests/clickpipes/stages/redpanda_user_data_mtls.sh.template
