--- crates/clickhouse-cloud-api/tests/integration/mod.rs
@@ -1,5 +1,0 @@
-#![allow(dead_code)]
-
-pub mod driver;
-pub mod stages;
-pub mod support;
