+++ crates/clickhouse-cloud-api/tests/integration_test.rs
@@ -1,8 +1,8 @@
-mod integration;
+mod common;
 
 use clickhouse_cloud_api::models::*;
 use clickhouse_cloud_api::Client;
-use integration::support::*;
+use common::support::*;
 
 #[tokio::test]
 #[ignore = "requires live ClickHouse Cloud credentials and provisions real resources"]
